+++ packages/stg-evm-v2/devtools/config/testnet/treasurer.config.ts
@@ -6,7 +6,7 @@ import { EndpointId } from '@layerzerolabs/lz-definitions'
 
 import { createGetAssetAddresses, getNamedAccount } from '../../../ts-src/utils/util'
 
-import { onArb, onBL3, onBsc, onEth, onKlaytn, onMantle, onOdyssey, onOpt } from './utils'
+import { onArb, onBL3, onBsc, onEth, onKlaytn, onMantle, onMonad, onOdyssey, onOpt } from './utils'
 
 const contract = { contractName: 'Treasurer' }
 const getDeployer = getNamedAccount('deployer')
@@ -22,6 +22,7 @@ export default async (): Promise<OmniGraphHardhat<TreasurerNodeConfig, unknown>>
     const bl3 = await getEnvironment(EndpointId.BL3_V2_TESTNET)
     const odyssey = await getEnvironment(EndpointId.ODYSSEY_V2_TESTNET)
     const mantle = await getEnvironment(EndpointId.MANTLESEP_V2_TESTNET)
+    const monad = await getEnvironment(EndpointId.MONAD_V2_TESTNET)
 
     // Then grab the deployer account for each network to be used as the admin
     const ethAdmin = await eth.getNamedAccounts().then(getDeployer)
@@ -32,6 +33,7 @@ export default async (): Promise<OmniGraphHardhat<TreasurerNodeConfig, unknown>>
     const bl3Admin = await bl3.getNamedAccounts().then(getDeployer)
     const odysseyAdmin = await odyssey.getNamedAccounts().then(getDeployer)
     const mantleAdmin = await mantle.getNamedAccounts().then(getDeployer)
+    const monadAdmin = await monad.getNamedAccounts().then(getDeployer)
 
     // Now we collect the address of the deployed assets
     const allAssets = [TokenName.USDT, TokenName.USDC, TokenName.ETH] as const
@@ -44,6 +46,7 @@ export default async (): Promise<OmniGraphHardhat<TreasurerNodeConfig, unknown>>
     const bl3AssetAddresses = await getAssetAddresses(EndpointId.BL3_V2_TESTNET, allAssets)
     const odysseyAssetAddresses = await getAssetAddresses(EndpointId.ODYSSEY_V2_TESTNET, allAssets)
     const mantleAssetAddresses = await getAssetAddresses(EndpointId.MANTLESEP_V2_TESTNET, allAssets)
+    const monadAssetAddresses = await getAssetAddresses(EndpointId.MONAD_V2_TESTNET, allAssets)
 
     return {
         contracts: [
@@ -130,6 +133,17 @@ export default async (): Promise<OmniGraphHardhat<TreasurerNodeConfig, unknown>>
                         [mantleAssetAddresses.USDT]: true,
                         [mantleAssetAddresses.USDC]: true,
                         [mantleAssetAddresses.ETH]: true,
+                    },
+                },
+            },
+            {
+                contract: onMonad(contract),
+                config: {
+                    admin: monadAdmin,
+                    assets: {
+                        [monadAssetAddresses.USDT]: true,
+                        [monadAssetAddresses.USDC]: true,
+                        [monadAssetAddresses.ETH]: true,
                     },
                 },
             },