+++ packages/stg-evm-v2/test/devtools/utils.test.ts
@@ -8,7 +8,10 @@ import { EndpointId } from '@layerzerolabs/lz-definitions'
 
 import {
     getContracts as getContractsMainnet,
+    getContracts as getContractsMainnet,
     isValidCreditMessagingChain as isValidCreditMessagingChainMainnet,
+    isValidCreditMessagingChain as isValidCreditMessagingChainMainnet,
+    validCreditMessagingChains as validCreditMessagingChainsMainnet,
     validCreditMessagingChains as validCreditMessagingChainsMainnet,
 } from '../../devtools/config/mainnet/utils'
 import {